requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic_core==2.33.1
pytesseract==0.3.13
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
